requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.0
openai==0.27.4
prometheus_client==0.16.0
google-genai